+++ include/game_start.h
@@ -1,0 +1,14 @@
+#ifndef POKEPLATINUM_GAME_START_H
+#define POKEPLATINUM_GAME_START_H
+
+#include "overlay_manager.h"
+#include "savedata.h"
+
+extern const OverlayManagerTemplate Unk_ov57_021D0F70;
+extern const OverlayManagerTemplate Unk_ov57_021D0F80;
+extern const OverlayManagerTemplate Unk_ov57_021D0F90;
+
+void ov57_021D0F30(int param0, SaveData *param1);
+void ov57_021D0F44(int param0, SaveData *param1);
+
+#endif // POKEPLATINUM_GAME_START_H
