--- include/overlay057/const_ov57_021D0F70.h
@@ -1,8 +1,0 @@
-#ifndef POKEPLATINUM_CONST_OV57_021D0F70_H
-#define POKEPLATINUM_CONST_OV57_021D0F70_H
-
-#include "overlay_manager.h"
-
-extern const OverlayManagerTemplate Unk_ov57_021D0F70;
-
-#endif // POKEPLATINUM_CONST_OV57_021D0F70_H
