--- include/overlay057/ov57_021D0F30.h
@@ -1,9 +1,0 @@
-#ifndef POKEPLATINUM_OV57_021D0F30_H
-#define POKEPLATINUM_OV57_021D0F30_H
-
-#include "savedata.h"
-
-void ov57_021D0F30(int param0, SaveData *param1);
-void ov57_021D0F44(int param0, SaveData *param1);
-
-#endif // POKEPLATINUM_OV57_021D0F30_H
