+++ platinum.us/sbins.sha1
@@ -56,7 +56,7 @@ b38da9fe733a6a1213ff217f4e2788e3ae2011a4 *overlay47.sbin
 395bbb8597227e687f61cea574c2219c7c77d1d1 *overlay54.sbin
 ea68f11d110d5929d18eeaad9cf3f1792986ab7f *overlay55.sbin
 1b89fbbf7576f21608df1334606ed2e77f5e21ea *overlay56.sbin
-88b06ba661770c8feef9ec96bcee6942819fbda4 *overlay57.sbin
+88b06ba661770c8feef9ec96bcee6942819fbda4 *game_start.sbin
 9b7c8c69b165a8a4f1da4ff4fc73bdc0af52def8 *overlay58.sbin
 43b9d869fc580cb7e6af6ef95cbbc015b5542d50 *overlay59.sbin
 0a0cfedfb93a08ff8fd187d6e62a5c9a1e3056cc *overlay60.sbin
