+++ src/meson.build
@@ -730,8 +730,7 @@ pokeplatinum_c = files(
     'overlay055/ov55_022561C0.c',
     'overlay055/ov55_0225632C.c',
     'overlay056/ov56_022561C0.c',
-    'overlay057/ov57_021D0D80.c',
-    'overlay057/ov57_021D0F30.c',
+    'game_start.c',
     'overlay058/ov58_021D0D80.c',
     'overlay059/ov59_021D0D80.c',
     'overlay059/ov59_021D2A2C.c',
